refactoring slicing_funcs
