+++ python-client/giskard/testing/tests/debug_slicing_functions.py
